feat: Connecting Posts and Users

+++ controllers/feed.js
@@ -2,6 +2,7 @@ const { validationResult } = require("express-validator");
 const Post = require("../models/post");
 const fs = require("fs");
 const path = require("path");
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -49,22 +50,32 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   const imageUrl = req.file.path;
+  let creator;
 
   // Create post in db
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Rishu" },
+    creator: req.userId
   });
 
   post
     .save()
     .then((result) => {
       console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: "Post created successfully!",
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch((err) => {
@@ -126,6 +137,12 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if(post.creator.toString() !== req.userId){
+        const error = new Error('Not Authorized');
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -163,13 +180,26 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
 
+      if(post.creator.toString() !== req.userId){
+        const error = new Error('Not Authorized');
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
 
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       console.log(result);
-      res.status(200).json({ message: "Deleted Post" });
+      res.status(200).json({ message: "Post Deleted" });
     })
     .catch((err) => {
       if (!err.statusCode) {
